requirements.txt
click==7.1.1
h11==0.9.0
jsonpickle==1.3
paho-mqtt==1.5.0
psutil==5.7.0
pydantic==1.4
starlette==0.13.2
websockets==8.1
